Fix illuminate 9 container test compatibility
